backend/requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
httpx
python-jose
passlib
slowapi
celery
redis
sqlalchemy
alembic
biopython
pyvcf3
pysam
torch
transformers
sentence-transformers
xgboost
scikit-learn
numpy
pandas
scipy
qdrant-client
langchain
langchain-community
langchain-ollama
reportlab
python-dotenv
cryptography
requests